#416 Move LED animations into single file headers
