+++ main/modes/utilities/dance/dance.h
@@ -10,35 +10,32 @@
 
 #include "swadge2024.h"
 
-typedef void (*ledDance)(uint32_t, uint32_t, bool);
+// Helper macros to pack an RGB color into an argument
+#define RGB_2_ARG(r, g, b) ((((r) & 0xFF) << 16) | (((g) & 0xFF) << 8) | (((b) & 0xFF)))
+#define ARG_R(arg)         (((arg) >> 16) & 0xFF)
+#define ARG_G(arg)         (((arg) >> 8) & 0xFF)
+#define ARG_B(arg)         (((arg) >> 0) & 0xFF)
+
+/**
+ * @brief A function to animate LEDs
+ *
+ * @param tElapsedUs The time since this was last called
+ * @param arg An optional argument such as color, timing, etc.
+ * @param reset true to reset state variables, false to run normally
+ */
+typedef void (*ledDanceFunc)(uint32_t tElapsedUs, uint32_t arg, bool reset);
 
 typedef struct
 {
-    ledDance func;
-    uint32_t arg;
-    char* name;
-} ledDanceArg;
+    ledDanceFunc func; ///< A 'main loop' function to animate and set the LEDs
+    uint32_t arg;      ///< An optional argument for the \c func, may be color, speed, etc.
+    char* name;        ///< The name of this LED dance
+} ledDance_t;
 
 extern swadgeMode_t danceMode;
-extern const ledDanceArg ledDances[];
-
-void danceComet(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceFlashlight(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceRise(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void dancePulse(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceSmoothRainbow(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceSharpRainbow(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceRainbowSolid(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceBinaryCounter(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceFire(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void dancePoliceSiren(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void dancePureRandom(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceRandomDance(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceChristmas(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceNone(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceCondiment(uint32_t tElapsedUs, uint32_t arg, bool reset);
-void danceSweep(uint32_t tElapsedUs, uint32_t arg, bool reset);
+extern const ledDance_t ledDances[];
 
 uint8_t getNumDances(void);
+uint32_t danceRand(uint32_t bound);
 
 #endif /* MODE_DANCE_H_ */
